type.tsを追加
