+++ src/components/Banner/index.ts
@@ -1,2 +1,2 @@
 export { default } from "./Banner";
-export type { BannerProps } from "./Banner";
+export type { BannerProps, BannerType, BannerSize } from "./types";
